+++ src/state/shareableStore/index.ts
@@ -1,1 +1,2 @@
 export * from './shareableStore'
+export * from './schemas'
